perf: cache tweets and load Sandpack client-only (prerender post)
post/[slug] had three client embeds blocking prerender. With all handled, the
route drops its page-level Suspense and prerenders.

- tweet: wrap getTweet() in "use cache" + cacheLife('max') (tweets are
  immutable) — fetched once and cached instead of hitting the syndication API
  per request
- sandbox: load Sandpack via dynamic(ssr: false) + Suspense with a skeleton
  (it reads Date.now() at render); split into sandbox-impl.tsx
- post/[slug]: drop the now-redundant page-level Suspense
